+++ src/components/articles/article_detail_connected_component.tsx
@@ -4,7 +4,7 @@ import { Store } from '../../stores';
 import { ArticleDetailComponent, ArticleDetailProps } from './article_detail';
 import { TaskListItemActions } from '../tasks/task_list_item';
 import { Id } from '../../stores/articles';
-import { selectArticle } from '../../selectors/articles';
+import { selectArticle } from '../../selectors/articles/select_article';
 import { AddToSavedListAction, addToSavedList } from '../../stores/tasks';
 import { RouterProps } from '../../application/routing';
 import { selectSavedTasksIdList } from '../../selectors/tasks';
