+++ src/selectors/articles/select_article.ts
@@ -1,0 +1,24 @@
+import { Store } from '../../stores';
+import * as taskDetails from '../details/tasks';
+import { selectLocale } from '../locale';
+import { selectRelatedTasks } from '../tasks';
+import { RouterProps } from '../../application/routing';
+import { buildExploreSection } from '../details/explore';
+import { selectExploreTaxonomy } from '../taxonomies';
+import { selectIconFromExploreTaxonomy } from '../select_icon_from_explore_taxonomy';
+import { toSelectorArticle } from './to_selector_article';
+import { toSelectorArticleList } from './to_selector_article_list';
+import { Article } from '.';
+
+export const selectArticle = (store: Store, routerProps: RouterProps): Article => {
+    const locale = selectLocale(store);
+    const articles = store.articlesInStore.articles;
+    const article = articles[routerProps.match.params.articleId];
+    const relatedTasks = selectRelatedTasks(store, article.relatedTasks);
+    const relatedArticles = toSelectorArticleList(store, article.relatedArticles);
+    const storeExploreSection = taskDetails.findExploreSectionBy(article, store.exploreSectionsInStore.sections);
+    const exploreTaxonomy = selectExploreTaxonomy(store);
+    const icon = selectIconFromExploreTaxonomy(storeExploreSection.taxonomyTerms, exploreTaxonomy);
+    const exploreSection = buildExploreSection(locale, storeExploreSection, icon);
+    return toSelectorArticle(locale, article, exploreSection, relatedArticles, relatedTasks);
+};
